refactor(asm): improve code quality based on PR feedback
