+++ crates/asm/common/src/lib.rs
@@ -8,7 +8,7 @@ mod msg;
 mod spec;
 mod state;
 mod subprotocol;
-mod subprotocol_genesis;
+mod genesis;
 mod tx;
 
 pub use aux::*;
@@ -18,7 +18,7 @@ pub use msg::*;
 pub use spec::*;
 pub use state::*;
 pub use subprotocol::*;
-pub use subprotocol_genesis::*;
+pub use genesis::*;
 pub use tx::*;
 
 // Re-export the logging module
