+++ crates/asm/logs/src/constants.rs
@@ -2,20 +2,18 @@ use strata_msg_fmt::TypeId;
 
 /// Macro to define all type IDs and ensure they're included in uniqueness tests
 macro_rules! define_ids {
-    ($type:ty, $fn_name:ident, $($name:ident = $value:expr),* $(,)?) => {
+    ($type:ty, $const_name:ident, $($name:ident = $value:expr),* $(,)?) => {
         $(
             pub const $name: $type = $value;
         )*
 
-        /// Get all defined type IDs as an array
-        pub const fn $fn_name() -> &'static [$type] {
-            &[$($name),*]
-        }
+        /// Array containing all defined type IDs
+        pub const $const_name: &'static [$type] = &[$($name),*];
     };
 }
 
 // Define all log type IDs
-define_ids! {TypeId, all_log_type_ids,
+define_ids! {TypeId, LOG_TYPE_IDS,
     DEPOSIT_LOG_TYPE_ID = 1,
     FORCED_INCLUSION_LOG_TYPE_ID = 2,
     CHECKPOINT_UPDATE_LOG_TYPE = 3,
@@ -32,7 +30,7 @@ mod tests {
 
     #[test]
     fn test_all_type_ids_are_unique() {
-        let log_ids = all_log_type_ids();
+        let log_ids = LOG_TYPE_IDS;
         let unique_ids: HashSet<_> = log_ids.iter().collect();
         assert_eq!(
             log_ids.len(),
