+++ crates/asm/subprotocols/bridge-v1/src/constants.rs
@@ -2,21 +2,19 @@ use strata_l1_txfmt::SubprotocolId;
 
 /// Macro to define all type IDs and ensure they're included in uniqueness tests
 macro_rules! define_ids {
-    ($type:ty, $fn_name:ident, $($name:ident = $value:expr),* $(,)?) => {
+    ($type:ty, $const_name:ident, $($name:ident = $value:expr),* $(,)?) => {
         $(
             pub(crate) const $name: $type = $value;
         )*
 
-        /// Get all defined type IDs as an array
+        /// Array containing all defined type IDs
         #[allow(dead_code)]
-        const fn $fn_name() -> &'static [$type] {
-            &[$($name),*]
-        }
+        const $const_name: &'static [$type] = &[$($name),*];
     };
 }
 
 // Define all subprotocol IDs
-define_ids! {SubprotocolId, all_subprotocol_ids,
+define_ids! {SubprotocolId, SUBPROTOCOL_IDS,
     CORE_SUBPROTOCOL_ID = 1,
     BRIDGE_SUBPROTOCOL_ID = 2,
 }
@@ -29,7 +27,7 @@ mod tests {
 
     #[test]
     fn test_all_type_ids_are_unique() {
-        let subprotocol_ids = all_subprotocol_ids();
+        let subprotocol_ids = SUBPROTOCOL_IDS;
         let unique_ids: HashSet<_> = subprotocol_ids.iter().collect();
         assert_eq!(
             subprotocol_ids.len(),
