+++ crates/asm/subprotocols/core/src/parsing/mod.rs
@@ -1,6 +1,6 @@
 //! This module handles checkpoint data extraction.
 
-pub(crate) mod checkpoint;
+mod checkpoint;
 
 // Re-export main parsing functions for convenience
 pub(crate) use checkpoint::extract_signed_checkpoint;
